moved header and horizontal_line to style module
