+++ lca_activity_browser/app/ui/tabs/history.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from PyQt5 import QtCore, QtWidgets
 
-from .. import horizontal_line, header
+from ..style import horizontal_line, header
 from ..tables import ActivitiesHistoryTable
