+++ lca_activity_browser/app/ui/tabs/lca_results.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from PyQt5 import QtWidgets
 
-from .. import horizontal_line, header
+from ..style import horizontal_line, header
 from ..tables import LCAResultsTable
 from ..graphics import (
     CorrelationPlot,
